improve argument handling and add style-attribute replacement
- use ArgumentParser to make argument handling easier
- make color, opacity and pngcrush optional
- remove fill and opacity properties from style Attribute if necessary
- remove the temp svg-file
- add .gitignore for .idea folder

+++ pngs_from_svg_xml.py
@@ -1,11 +1,12 @@
 #!/usr/bin/python3
 
-import sys
 import os
 import shlex
 import subprocess
 import tempfile
 import shutil
+import argparse
+import re
 
 import xml.etree.ElementTree as ElementTree
 
@@ -20,17 +21,28 @@ def modify_svg(svg, tmp, color, opacity):
     root = tree.getroot()
 
     for path in root.findall("{" + ns + "}path"):
-        print(path)
         fill = path.get('fill')
-        if fill is None:
-            print(fill)
-            path.set("fill", color)
-            path.set("opacity", str(opacity))
+        if fill != "none":
+            style = path.get('style')
+
+            if color:
+                if style is not None:
+                    style = re.sub("fill:\s*#\d{3,6};?", "", style)
+                    path.set("style", style)
+
+                path.set("fill", color)
+
+            if opacity:
+                if style is not None:
+                    style = re.sub("opacity:\s*\d?\.?\d*;?", "", style)
+                    path.set("style", style)
+
+                path.set("opacity", str(opacity))
 
     tree.write(tmp)
 
 
-def create_images(svg, dir_pngs, namebase, suffix, isize, color, opacity):
+def create_images(svg, dir_pngs, namebase, suffix, isize, color, opacity, nopngcrush=False):
     # dimensions for mdpi/hdpi/xhdpi/xxhdpi
     sizes = [isize, isize * 1.5, isize * 2, isize * 3]
 
@@ -49,27 +61,27 @@ def create_images(svg, dir_pngs, namebase, suffix, isize, color, opacity):
         output = os.path.join(output_dir, namebase + suffix + ".png")
         print (svg + ' -> ' + output)
 
-        create_png(tmp_svg[1], output, size)
+        create_png(tmp_svg[1], output, size, nopngcrush)
 
-    #os.remove(tmp_svg[1])
+    os.remove(tmp_svg[1])
 
 
-def create_png(svg, output, size):
+def create_png(svg, output, size, nopngcrush):
     cmd = "inkscape -C -e " + output + " -h " + str(size) + " " + svg
     args = shlex.split(cmd, False, running_on_posix)
     subprocess.call(args)
 
-    tmp_png = tempfile.mkstemp(".png")
-    os.close(tmp_png[0])
-    shutil.copyfile(output, tmp_png[1])
-    os.remove(output)
+    if nopngcrush == False:
+        tmp_png = tempfile.mkstemp(".png")
+        os.close(tmp_png[0])
+        shutil.copyfile(output, tmp_png[1])
+        os.remove(output)
 
-    cmd = "pngcrush " + tmp_png[1] + " " + output
-    args = shlex.split(cmd, False, running_on_posix)
-    print (args)
-    subprocess.call(args)
+        cmd = "pngcrush " + tmp_png[1] + " " + output
+        args = shlex.split(cmd, False, running_on_posix)
+        subprocess.call(args)
 
-    os.remove(tmp_png[1])
+        os.remove(tmp_png[1])
 
 
 def is_posix():
@@ -83,25 +95,19 @@ running_on_posix = is_posix()
 
 if __name__ == "__main__":
 
-    nargs = len(sys.argv)
-    if nargs != 6 and nargs != 7:
-        print ("usage: " + sys.argv[0] +
-               " <filename (input)> <res directory (output)> <size> <color> <opacity> [<suffix>]")
-        exit(1)
+    parser = argparse.ArgumentParser()
+    parser.add_argument("input_filename",      help="Path to the svg file")
+    parser.add_argument("output_directory",    help="Output directory")
+    parser.add_argument("size",                help="Size in pixel for the mdpi variant", type=int)
+    parser.add_argument("-c", "--color",       help="Color, e.g. '#104E8B'")
+    parser.add_argument("-o", "--opacity",     help="Opacity, e.g. '0.54'",               type=float)
+    parser.add_argument("-s", "--suffix",      help="Suffix for the output filenames",    default="")
+    parser.add_argument("-nc", "--nopngcrush", help="Don't run pngcrush", action="store_true")
+    args = parser.parse_args()
 
-    svg_path = sys.argv[1]
-    dir_pngs = sys.argv[2]
-    isize = int(sys.argv[3])
-    color = sys.argv[4]
-    opacity = float(sys.argv[5])
-
-    suffix = ""
-    if nargs == 7:
-        suffix = sys.argv[6]
-
-    svg_name = os.path.basename(svg_path)
+    svg_name = os.path.basename(args.input_filename)
     name = svg_name
     if svg_name.lower().endswith(".svg"):
         name = svg_name[:-4]
 
-    create_images(svg_path, dir_pngs, name, suffix, isize, color, opacity)
+    create_images(args.input_filename, args.output_directory, name, args.suffix, args.size, args.color, args.opacity, nopngcrush=args.nopngcrush)
